requirements.txt
opencv==4.2.0
flask==1.1.2
imutils==0.5.3
Pillow==2.2.2
pytesseract==0.3.4
scikit-image==0.17.2
Keras==2.3.1
tensorflow==2.0.0

